+++ core/src/main/java/co/ericp/freewayforecast/routes/Route.kt
@@ -6,9 +6,9 @@ import co.ericp.freewayforecast.Location
  * A route at a particular time from a place to a place.
  */
 class Route(
-        val distance: Long, // Meters
-        val duration: Long, // Unix time
-        val startTime: Long,
+        val distance: Double, // Meters
+        val duration: Long, // Milliseconds
+        val startTime: Long, // Unix time (millis)
         val endTime: Long,
         val neBound: Location,
         val swBound: Location,
